Refactored plugins
This is a huge change. Plug wasn’t best fit for us, because it decides
which modules and functions should be executed at compile time. Instead
we need to preprocess request, decide which plugins are enabled and
they execute them.

Additional features are:
- Request features set will allow us to do smart optimizations on a
request handling. Eg. we can write log asynchronously when idempotency
plugins is disabled. Or don’t decode request body if validation is not
needed.
- First step in moving from Plug.Router to a standalone proxy module,
which will allow to proxy file uploads without persisting them to disk.